requirements.txt
Pillow==10.3.0
numpy==1.26.4
tqdm==4.66.2
scikit-learn==1.2.0
scikit-image==0.23.2
ultralytics==8.0.220
customtkinter==5.2.2
opencv-python==4.9.0.80
pandas==2.2.2.
matplotlib==3.8.4







